WIP trial of Content-Security-Policy headers
The goal is to allow rapid experimentation until we have a reasonably
useful CSP header for at least on large IDA, and then to move this code
to a more appropriate location. It is specifically under cms/ to avoid
the temptation to use it on the LMS side before that move happens.